+++ src/components/common/TimePicker/styles.css
@@ -1,0 +1,15 @@
+.swiper {
+  width: 100%;
+  height: 100%;
+}
+
+.swiper-slide {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: #c7ccd6;
+}
+
+.swiper-slide-active {
+  color: #646e81;
+}
